Try to build an openpose Docker image
Building the image (compiling openpose) was very long, 1 hour.
I do not have a GPU on my image and it seems complex see
run_openpose.md to see CUDA warnings and support-policy.
We should use a base image that depends on the GPU hardware of the
machine (their drivers) that is still maintained.
I don't know if we can build an image on one machine to be deployed on
another machine.